Onboard combined to single route
