+++ src/docs/components/DocsAppBar/DocsAppBar.svelte
@@ -206,7 +206,7 @@
 		</nav>
 		<hr class="my-4" />
 		<div>
-			<a class="btn variant-ghost-surface w-full" href="/guides/themes/generator">Theme Generator</a>
+			<a class="btn variant-ghost-surface w-full" href="/docs/get-started/generator">Theme Generator</a>
 		</div>
 	</div>
 </AppBar>
