+++ src/routes/(inner)/docs/get-started/colors/+page.svelte
