+++ src/routes/(inner)/docs/get-started/generator/+page.svelte
