+++ src/routes/(inner)/docs/quickstart/+page.svelte
